Added List Usage

+++ Project_1/utils/utils.hpp
@@ -8,6 +8,6 @@
 
 inline uint32_t BitConversion (uint64_t, int);
 void            SwitchElements(uint64_t **, int, int, int);
-void            MergeTables   (List<int> &, uint64_t **, uint32_t *, int, int, uint64_t **, uint32_t *, int, int);
+void            MergeTables   (List<uint64_t> &, uint64_t **, uint32_t *, int, int, uint64_t **, uint32_t *, int, int);
 
 #endif // __UTILS_HEADER__
